# db/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.29.0

package dbgen

import (
	"time"

	"github.com/jackc/pgx/v5/pgtype"
)

type Cart struct {
	ID          int64     `json:"id"`
	ProductName string    `json:"product_name"`
	Qty         int32     `json:"qty"`
	Buyer       string    `json:"buyer"`
	Price       float64   `json:"price"`
	OrderID     int64     `json:"order_id"`
	CreatedAt   time.Time `json:"created_at"`
	UpdatedAt   time.Time `json:"updated_at"`
}

type Order struct {
	ID          int64   `json:"id"`
	OrderNumber string  `json:"order_number"`
	Discount    float64 `json:"discount"`
	// pending, completed
	Status    string    `json:"status"`
	Total     float64   `json:"total"`
	CreatedAt time.Time `json:"created_at"`
	UpdatedAt time.Time `json:"updated_at"`
}

type Payment struct {
	ID            int64  `json:"id"`
	PaymentNumber string `json:"payment_number"`
	OrderID       int64  `json:"order_id"`
	// bank, e-wallet, etc
	Method    string             `json:"method"`
	Total     float64            `json:"total"`
	PaidAt    pgtype.Timestamptz `json:"paid_at"`
	CreatedAt time.Time          `json:"created_at"`
	UpdatedAt time.Time          `json:"updated_at"`
}